Add openPersonCard and hidePid attributes.
hidePid is needed because inline people sometimes have a pid for opening person card, but we don't want to show it.

_openPersonCard() checks to see if the attribute is set and the person card exists on the page before triggering an event. The event listener and handler is in person-card.html. Although using a global event isn't the greatest way of doing things, it allows us to open a single instance of person card from anywhere without having to import it as a dependency (which would create a circular dependency if added here).

As a side note, we want to keep the destination href with openPersonCard, so that the href can be opened instead of the person card when a cmd + click or ctrl + click is used on the name.