OFT-69102 Fix copy pid functionality
The functionality to copy the PID is fixed as part of this. For some reason the width of the hidden input was causing the problem for person-card. I've tested it out and increasing the width of the hidden input seems to fix the issue.

There's also a change to make it so the PID itself will copy when clicked, so the tooltip has to show when hovering over that as well. This was a ruling made by Jeff Hawkins.
I've also added a unit test for making sure the aria-label gets added on copy and I added a little variety to the demo page since all of the person renderers had the same PID.